requirements.txt
h5py==2.9.0
numpy==1.16.4
PyYAML==5.1.1
six==1.12.0
